+++ react/__mocks__/vtex.render-runtime.tsx
@@ -1,0 +1,6 @@
+export const useRuntime = () => {
+  return {
+    culture: 'BRA',
+    account: '',
+  }
+}