+++ react/__mocks__/vtex.styleguide.tsx
@@ -1,0 +1,6 @@
+import Button from '@vtex/styleguide/lib/Button'
+import Input from '@vtex/styleguide/lib/Input'
+import Tag from '@vtex/styleguide/lib/Tag'
+import Radio from '@vtex/styleguide/lib/Radio'
+
+export { Button, Input, Tag, Radio }