+++ react/components/PostalCode.tsx
@@ -2,7 +2,9 @@ import React, { FunctionComponent, Fragment } from 'react'
 import { components } from 'vtex.address-form'
 import { modules } from 'vtex.country-codes'
 import { injectIntl, InjectedIntlProps, defineMessages } from 'react-intl'
+
 // @ts-ignore
+// eslint-disable-next-line
 const { toAlpha2 } = modules
 const { CountrySelector, PostalCodeGetter, StyleguideButton } = components
 
@@ -15,7 +17,7 @@ defineMessages({
 
 interface CustomProps {
   countries: string[]
-  handleSubmit: any
+  handleSubmit: () => void
 }
 
 type Props = CustomProps & InjectedIntlProps