+++ CHANGELOG.md
@@ -7,6 +7,15 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## Added
+
+- Selects a delivery option.
+- Estimates shipping value.
+
+## Changed
+
+- Title and postal code margin values.
+
 ## [0.2.1] - 2019-08-23
 
 ## Fixed